Former-commit-id: f625ca79deecaabaf625ac80efe85db15cd207bc

+++ cmake/generateVersionHeader.py
@@ -18,9 +18,8 @@ commit = os.popen('git rev-parse --verify HEAD' ).read().rstrip("\n")
 #print commit
 #print commit[0:8]
 
-tag = os.popen( 'git describe --tags $(git rev-list --tags --max-count=1)' ).read().rstrip("\n")
-
-#print tag
+tag = os.popen( 'git describe --tags HEAD' ).read().rstrip("\n")
+tag = tag[0:tag.find("-")]
 
 command = "git log -1 --format=%ai " + str( tag )
 datetime = os.popen( command ).read().rstrip("\n")
@@ -76,7 +75,7 @@ f.write("#define REST_VERSION(a,b,c) (((a) << 16) + ((b) << 8) + (c))\n")
 f.write("#endif\n")
 f.close()
 
-if filecmp.cmp( "TRestVersion.tmp", outputHeader ):
+if os.path.exists( outputHeader ) and filecmp.cmp( "TRestVersion.tmp", outputHeader ):
     os.remove( "TRestVersion.tmp" )
 else:
     os.rename( "TRestVersion.tmp", outputHeader )
